Move ref management to dedicated hook
